+++ nvim/lua/custom/settings.lua
@@ -1,0 +1,13 @@
+local M = {
+  vault = {
+    name = 'default',
+    path = '~/vaults/default',
+    inbox_dir = 'inbox',
+    daily_dir = 'daily',
+    notes_dir = 'notes',
+    attachments_dir = 'assets',
+    templates_dir = 'templates',
+  },
+}
+
+return M